Implement proc macro for sync::Trace
