+++ tracing-rc-derive/Cargo.toml
@@ -9,6 +9,9 @@ repository = "https://github.com/Jesterhearts/tracing-rc"
 [lib]
 proc-macro = true
 
+[features]
+sync = []
+
 [dependencies]
 proc-macro2 = '^1.0'
 quote = '^1.0'
