+++ tracing-rc-derive/src/lib.rs
@@ -2,33 +2,20 @@
 //! Procedural macro for deriving a `Trace` implementation for the `tracing_rc` crate.
 
 use proc_macro2::TokenStream;
-use quote::quote;
 
 const TRACE_FIELD_IDENT: &str = "trace";
 const IGNORE_FIELD_IDENT: &str = "ignore";
 
-fn trace_derive(mut s: synstructure::Structure) -> TokenStream {
+mod rc;
+#[cfg(feature = "sync")]
+mod sync;
+
+fn filter_default_and_ignored_fields(
+    default_fields: &mut synstructure::Structure,
+) -> Vec<TokenStream> {
     let mut errors = vec![];
 
-    let mut has_specific_field = false;
-
-    // All fields marked #[trace] but not #[trace(ignore)]
-    let mut specific_fields = s.clone();
-    let specific_fields = specific_fields.filter(|bi| {
-        if bi.ast().attrs.iter().any(|attr| {
-            attr.path.is_ident(TRACE_FIELD_IDENT) && attr.parse_args::<syn::Meta>().is_err()
-        }) {
-            has_specific_field = true;
-            true
-        } else {
-            false
-        }
-    });
-
-    let mut has_ignored = false;
-
-    // All fields excluding #[trace(ignore)]
-    let default_fields = s.filter(|bi| {
+    default_fields.filter(|bi| {
         let mut seen_attr = false;
         let mut should_keep = true;
 
@@ -44,8 +31,6 @@ fn trace_derive(mut s: synstructure::Structure) -> TokenStream {
                                 ).to_compile_error(),
                             );
                         }
-
-                        has_ignored = true;
                         false
                     }
                     _ => {
@@ -78,40 +63,23 @@ fn trace_derive(mut s: synstructure::Structure) -> TokenStream {
         should_keep
     });
 
-    let (body, s) = if has_specific_field {
-        (
-            specific_fields.each(|bi| {
-                quote! {
-                    tracing_rc::rc::Trace::visit_children(#bi, visitor);
-                }
-            }),
-            specific_fields,
-        )
-    } else {
-        (
-            default_fields.each(|bi| {
-                quote! {
-                    tracing_rc::rc::Trace::visit_children(#bi, visitor);
-                }
-            }),
-            default_fields,
-        )
-    };
+    errors
+}
 
-    let trace_impl = s.underscore_const(true).unbound_impl(
-        quote!(tracing_rc::rc::Trace),
-        quote! {
-                fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                    match *self { #body }
-                }
-        },
-    );
+fn filter_specific_fields(specific_fields: &mut synstructure::Structure) -> bool {
+    let mut has_specific_field = false;
+    specific_fields.filter(|bi| {
+        if bi.ast().attrs.iter().any(|attr| {
+            attr.path.is_ident(TRACE_FIELD_IDENT) && attr.parse_args::<syn::Meta>().is_err()
+        }) {
+            has_specific_field = true;
+            true
+        } else {
+            false
+        }
+    });
 
-    quote! {
-        #trace_impl
-
-        #(#errors)*
-    }
+    has_specific_field
 }
 
 synstructure::decl_derive! {
@@ -158,240 +126,53 @@ synstructure::decl_derive! {
     /// }
     /// ```
     ///
-    trace_derive
+    rc::trace_derive
 }
 
-#[test]
-fn simple_impls_trace_default_all() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            struct Simple {
-                a: tracing_rc::rc::Gc<i64>,
-
-                b: i64,
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl tracing_rc::rc::Trace for Simple {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            Simple {
-                                a: ref __binding_0,
-                                b: ref __binding_1,
-                            } => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                                { tracing_rc::rc::Trace::visit_children(__binding_1, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
-}
-
-#[test]
-fn simple_impls_trace_specific_field() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            struct Simple {
-                #[trace]
-                a: tracing_rc::rc::Gc<i64>,
-
-                b: i64,
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl tracing_rc::rc::Trace for Simple {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            Simple {
-                                a: ref __binding_0,
-                                ..
-                            } => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
-}
-
-#[test]
-fn simple_impls_trace_specific_field_ignore() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            struct Simple {
-                a: tracing_rc::rc::Gc<i64>,
-
-                #[trace(ignore)]
-                b: i64,
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl tracing_rc::rc::Trace for Simple {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            Simple {
-                                a: ref __binding_0,
-                                ..
-                            } => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
-}
-
-#[test]
-fn enum_impls_trace() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            enum AnEnum {
-                None,
-                Some(#[trace] tracing_rc::rc::Gc<i64>),
-                StructLike {
-                    non_gc: i64,
-                    #[trace]
-                    gc_field: tracing_rc::rc::Gc<i64>,
-                },
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl tracing_rc::rc::Trace for AnEnum {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            AnEnum::None => {}
-                            AnEnum::Some(ref __binding_0,) => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                            }
-                            AnEnum::StructLike{gc_field: ref __binding_1, ..} => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_1, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
-}
-
-#[test]
-fn generics_impl_trace_default_all() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            struct HasGenerics<T: Trace, U: SomeTrait> {
-                a: Gc<T>,
-                b: U,
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl<T: Trace, U: SomeTrait> tracing_rc::rc::Trace for HasGenerics<T, U> {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            HasGenerics{
-                                a: ref __binding_0,
-                                b: ref __binding_1,
-                            } => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                                { tracing_rc::rc::Trace::visit_children(__binding_1, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
-}
-
-#[test]
-fn generics_impl_trace_default_specific() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            struct HasGenerics<T: Trace, U: SomeTrait> {
-                #[trace]
-                gc_field: Gc<T>,
-
-                non_gc: U,
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl<T: Trace, U: SomeTrait> tracing_rc::rc::Trace for HasGenerics<T, U> {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            HasGenerics{gc_field: ref __binding_0, ..} => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
-}
-
-#[test]
-fn generics_impl_trace_default_specific_ignore() {
-    synstructure::test_derive! {
-        trace_derive {
-            #[derive(Trace)]
-            struct HasGenerics<T: Trace, U: SomeTrait> {
-                gc_field: Gc<T>,
-
-                #[trace(ignore)]
-                non_gc: U,
-            }
-        }
-        expands to {
-            const _: () = {
-                extern crate tracing_rc;
-
-                impl<T: Trace, U: SomeTrait> tracing_rc::rc::Trace for HasGenerics<T, U> {
-                    fn visit_children(&self, visitor: &mut tracing_rc::rc::GcVisitor) {
-                        match *self {
-                            HasGenerics{gc_field: ref __binding_0, ..} => {
-                                { tracing_rc::rc::Trace::visit_children(__binding_0, visitor); }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-        no_build
-    };
+#[cfg(feature = "sync")]
+synstructure::decl_derive! {
+    [SyncTrace, attributes(trace)] =>
+    /// By default, this will call `Trace::visit_children` on all members of the type.
+    /// - If you wish to only visit specific members, you may annotate them with `#[trace]`, and
+    ///   only those members will be traced (other fields will be ignored).
+    /// - If you wish to _ignore_ a specific field, you may annote it with `#[trace(ignore)]`.
+    ///
+    /// # Examples
+    /// - Default:
+    /// ```
+    /// # use tracing_rc_derive::SyncTrace;
+    /// # use tracing_rc::sync::Agc;
+    /// #[derive(SyncTrace)]
+    /// struct MyType {
+    ///     default_traced_1: Agc<u64>,
+    ///     default_traced_2: Agc<u64>,
+    /// }
+    /// ```
+    /// - Tracing specific fields:
+    /// ```
+    /// # use tracing_rc_derive::SyncTrace;
+    /// # use tracing_rc::sync::Agc;
+    /// #[derive(SyncTrace)]
+    /// struct MyType {
+    ///     #[trace]
+    ///     gc_field: Agc<u64>,
+    ///
+    ///     default_ignored: i64,
+    /// }
+    /// ```
+    ///
+    /// - Ignoring specific fields:
+    /// ```
+    /// # use tracing_rc_derive::SyncTrace;
+    /// # use tracing_rc::sync::Agc;
+    /// #[derive(SyncTrace)]
+    /// struct MyType {
+    ///     default_traced: Agc<u64>,
+    ///
+    ///     #[trace(ignore)]
+    ///     ignored: i64,
+    /// }
+    /// ```
+    ///
+    sync::trace_derive
 }
