+++ tracing-rc/Cargo.toml
@@ -10,7 +10,8 @@ readme = "../README.md"
 [features]
 default = [ "proc_macro" ]
 proc_macro = [ "tracing-rc-derive" ]
-sync = [ "atomic", "once_cell", "dashmap", "parking_lot" ]
+sync = [ "sync_no_proc_macro", "tracing-rc-derive/sync" ]
+sync_no_proc_macro = [ "atomic", "once_cell", "dashmap", "parking_lot" ]
 
 [dependencies]
 indexmap = '^1.0'
